requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.1.3
drf-yasg==1.21.7
epicbox==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
structlog==23.1.0
uritemplate==4.1.1
urllib3==1.24.3
websocket-client==1.6.1
